#1295 Moved LFRicDimension to LFRicTypes.
